Conflicts:
	Airship/Common/UAAction.h
	Airship/Common/UAAction.m
	Airship/Common/UAActionArguments.h
	Airship/Common/UAActionArguments.m
	Airship/Common/UAActionRegistrar.h
	Airship/Common/UAActionRunner.h
	Airship/Common/UAActionRunner.m
	AirshipLib/AirshipLib.xcodeproj/project.pbxproj

+++ Airship/Common/UAAnalyticsDBManager.h
@@ -40,7 +40,7 @@ SINGLETON_INTERFACE(UAAnalyticsDBManager);
 - (void)resetDB;
 
 - (void)addEvent:(UAEvent *)event withSession:(NSDictionary *)session;
-- (NSArray *)getEvents:(int)max;
+- (NSArray *)getEvents:(NSUInteger)max;
 - (NSArray *)getEventByEventId:(NSString *)event_id;
 
 - (void)deleteEvent:(NSNumber *)eventId;
