+++ AirshipLib/AirshipLogicTests/UAHTTPRequestEngineTest.m
@@ -78,7 +78,7 @@
         [(UAHTTPConnection *)operation start];
     } else if ([operation isKindOfClass:[NSBlockOperation class]]) {
         NSInteger seconds = ((UADelayOperation *)operation).seconds;
-        NSLog(@"quote unquote sleeping for %d seconds", seconds);
+        NSLog(@"quote unquote sleeping for %ld seconds", (long)seconds);
     } else {
         XCTFail(@"got an unexpected operation type: %@", operation);
     }
